requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
PyPDF2>=3.0.0
requests>=2.28.0
